@@ -101,9 +101,9 @@
     "morgan": "^1.7.0",
     "ms": "^0.7.1",
     "nconf": "^0.8.4",
-    "react": "^0.14.8",
+    "react": "15.6.1",
     "react-codemirror": "^0.2.6",
-    "react-dom": "^0.14.8",
+    "react-dom": "15.6.1",
     "react-redux": "^4.4.5",
     "react-router": "^2.8.1",
     "react-router-redux": "^4.0.6",
